Alternate RNG for ExtraP Solution modification

+++ SimulatedAnnealingExtraP/LatexPrinter.cpp
@@ -266,7 +266,7 @@ void LatexPrinter<SolType>::printColorDefinitions(int number, ofstream & stream)
 			+ std::to_string(rand() % 255) + "}");
 	}
 
-	for (int i = 0; i < colors.size(); i++) {
+	for (size_t i = 0; i < colors.size(); i++) {
 		stream << colors[i].c_str() << std::endl;
 	}
